Apply a change that Sue requesed.
I would like to make one more logic tweak to the deapr program involving
the calculation of the Fold Chg, and the Abs Diff.
Hopefully it's a pretty simple change.  If the gene is a DELV gene,
leave the calculations just as they are.  But, if it is just an SRMM gene,
please populate the Fold Chg field with the SRMM calculation
(which should automatically update the Abs Fold Chg value as well),
and then calculate the Abs Diff by taking the min value of the
higher group and subtracting the max value of the lower group.
This should result in a change in the rankings, which will push
some of the more erratically expressed genes further down on the list.